Fix type mismatch with strtoul and strstr
Clang (and the standard) require that the return value of strstr be a
const char * if the first argument is const, and char * if it is not.
However, strtoul is not updated from the C definition in a similar way,
causing a type mismatch. This commit resolves that mismatch by making p
const and adding an additional endptr variable for the out parameter of
strtoul.