Address issues in ResourceIdentifier parsing and tests.
ResourceIdentifier parsing code overly optimistically assumed a fixed
length of the actual ResourceID.  The parsing code only checked that
the buffer length was larger than the resource prefix, and after that,
it assumed the actual resource ID was a fixed string of 16 hexadecimal
digits. This pull request addresses that shortcoming.

The test code was incorrect as well, passing the size of the pointer
into the `FromString` methods as opposed to size of the buffer.  This
led to the tests succeeding on platforms with 64 bit pointers (only in
conjunction with the bug above) and failing on platforms with 32 bit
pointers.