On sockets platforms, starting a timer should wake up layer's event loop
Calling `nl::Weave::System::Timer::Start` does not wake the Weave
thread. This is arguably an understood limitation stemming from Weave
code being geared towards a single-threaded execution.  However, on a
number of platforms -- such as Java -- we admit to having pieces of
the Weave stack run within different threads but enclosed within the
correct lock around the Weave stack.  This results in issues where the
Java caller schedules a timer: the main Weave thread may be sleeping
in the event loop and does not adjust the wake time in response to the
new timer being scheduled.

The construction proposed herein parallels the semantics
`nl::Weave::System::Timer::Start` offers on LwIP systems and parallels
the semantics offered by `ScheduleWork`.

On systems where POSIX locking is already enabled, calling
`WakeSelect` from the layer's event loop does not actually involve a
write to a pipe, the wake is only invoked when the `Timer::Start` is
invoked from a different thread.