Improve the search for OpenSSL on MacOS
MacOS migrated to LibreSSL sometime ago, and the preferred
way to use OpenSSL on MacOS is via `brew`.  This commit alters
the search algorithm for the preferred openssl executable to
first check the directory where brew actually installed the
executable; this addresses dynamic link issues with some partial
openssl installations.