Improved logging in python
-- Reworked the way developer logging happens when using the openweave wrappers for python.
With this change, all WeaveLog...() calls in C++ code result in a callback to a logging
function in python.  The default implementation of this function directs log messages to a
standard python logger object associated with the python WeaveStack object.  By default, the
python logger is configured to output messages to stdout, in a format identical to that used
by the C++ code.  However, developers are free to reconfigure the logger object, or replace
it altogether, giving them complete control over the handling of C++ generated logs in
python applications.

-- Added support for dynamic logging behavior.  Modified the standard implementation of
nl::Weave::Logging::Log() so that the default logging behavior can be overridden at runtime.
Applications can now use the SetLogFunct() function to specify a pointer to a function to be
called for every logged message.  Setting the function pointer to NULL (the default value)
results in the original behavior of logging to stdout via printf().