DeviceLayer and DeviceManager: Add system managed network
The system managed network uses the network already existing on the
system. It can be used for demonstration and daily development. It
requires no extra hardware at all, and easy to set up.

The patch includes interface to implement system managed network, it can
be implemented in the device layer of the target platform outside the
source tree. Linux device layer will have its own implementation using
the given interface.

Change-Id: Ibb488010e9fc2711bc0c00beba73d6e829a91fe5