DeviceManager: don't listen on 11095 when ephemeral is enabled
When both device emulator and device manager running on the same
station, both of them will listen on 11095, only one of them choosing
randomly can receive packets from the socket.

Ephemeral port is introduced to solve the problem, let device manager to
choose a random port instead of 11095. But there is a bug that even with
ephemeral enabled, device manager still listening on 11095.

The patch solved the bug, and prevent device manager from listening of
11095 when ephemeral is enabled.

Change-Id: I61e1b92fd329991d5931ce8472fa91580794bbcc