Tunnel CASE Session persistence support.
Changes within OpenWeave to support persistence of CASE session
for the Tunnel.
During connection establishment by the Tunnel Agent, if a persisted
session key exists, the TunnelAgent allows for only the WeaveConnection
to be established and then calls back into the application to load the
session key for the created Weave connection object.
When the connection goes down, and before the FabricState removes the
SessionKey, the application is called back to suspend the session into
a serialized form to be persisted by the application.
At time of suspension, resumption message Ids(for both the next message
id to send as well as the last received message id from peer) are
calculated using a fixed offset and included within the session key.
When the session resumes, it would start off communication using the
resumption message ids.

Add Test Client and Server for testing persistence of Tunnel CASE
session.