Fix incorrect usage of xEventGroupSetBitsFromISR
There were a few changes here:

1) xEventGroupSetBitsFromISR returns a BaseType_t but this return
value was being assigned to an EventBits_t and then being returned
from vRaiseEventFlagBasedOnContext.  Whatever values ended up in the
BaseType_t are not likely to be correct for an EventBits_t.  The fix
here is to change vRaiseEventFlagBasedOnContext to return void, since
none of the callers actually use the return value anyway.

2) Per the documentation at
https://www.freertos.org/xEventGroupSetBitsFromISR.html
xEventGroupSetBitsFromISR requires that portYIELD_FROM_ISR or
portEND_SWITCHING_ISR be called if xEventGroupSetBitsFromISR doesn't
fail.  This code was not doing that, and the fix was to call whichever
of those we have available.

3) The higherPrioTaskWoken outparam was unused by consumers and is
really only needed for the portYIELD_FROM_ISR/portEND_SWITCHING_ISR
call, so that got removed as well.