Add APIs to persist and restore state of subscriptions
This commit enables a set of APIs to persist and restore the
subscription state.  Such functionality may be used when the
Weave process needs to shutdown and restart in an orderly fashion
while preserving the illusion that the subscription remains
unchanged. In order to successfully accomplish this task, the
commit introduces the functionality to:

* serialize SubscriptionClient and SubscriptionHandler data.

* load SubscriptionClient and SubscriptionHandler from
  serialized data and integrate it into the subscription engine
  mechanisms.