Openweave Python packages should not depend on Python ABI version
We've noticed that Openweave Python packages on PyPI are only uploaded
for specific Python versions. This was a problem when we switched from
Python 3.7 to Python 3.8, and it was finally solved by uploading a new
Python package to PyPI. But now we've switched to Python 3.9 and the
problem surfaced again... so it's high time for a better solution.

Since Openweave Python package includes, and wraps around, a binary
shared object, it is not a "pure" package.  However, in that case,
Python's "wheel" packer assumes the package uses Python ABI (e.g. it
builds a Python module as a shared object), and tags it with the version
of ABI and the interpreter. Openweave does not use Python ABI, so this
is a completely unnecessary constraint. This PR removes it.