Make double sure the protocol callback is set on the binding object in SubscriptionClient
SubscriptionClient is written in a way that it's never released. When peer send a
"cancel subscription" request, SubscriptionClient is reset to initated state and
is going to be reused by the next subscription attempt. However, SubscriptionClient
only sets the protocol callback on its binding upon init(), which doesn't happen
when the same subscriptionClient is reused for the second (and new) subscription
attempt.
In some recent tests, we are oberving that binding is reset during AbortSubcription. When
subscriptionClient is reused for second subcription, protocol callback is not set
for its binding.  This PR is a temporary fix to unblock cancel/re-subscribe.
We should consider release subscriptionClient upon "cancel" to align with this: https://github.com/openweave/openweave-core/blob/9f9d2450ba3308f16cc729324801992d37130813/src/lib/profiles/data-management/Current/SubscriptionClient.h#L446