Do not change protocol callback when loading subscriptionHandler
In subscription persistence feature, we didn't persist binding, and
shared same binding between client and handler for mutual subscription.
The handler is only registered on the binding after the binding has been
prepared. The only meaningful event in the callback BindingFailed.
We should use the same protocol callback with subscriptionClient.