Weave Message Capture logic.
This feature allows an application to formulate an Exchange for
capturing Weave messages by setting a WeaveBindiing flag before vending
the ExchangeContext. For tunneled messages, they are directed towards a
'capture subnet' as a means of tagging the messages for capturing, as
they traverse the underlying network stack and get encapsulated within
an [IPv6|UDP] packet. The capture subnet is then reinstated back to the
original subnet before the packet is captured and delivered up to
the application.

* Expose WeaveBinding setting API.
* Percolate setting via message flags from WeaveExchange layer to
  WeaveMessageLayer.
* Encode reserved capture subnet for Tunneled packets to Phoenix
* Reinstate Phoenix address and tag tunneled packet for capture.
* Add functional test case for capturing a tunneled packet.